Two Lightweight JQuery Plugins. The first is a pie chart and the other is a circular progress bar.
Two Lightweight JQuery Plugins.
The first is a pie chart that can display statistics for several items
in a pie chart view.
And the other is a circular progress bar that show the progress in
percentage for a certain item in a circular bar instead of the normal
horizontal bar.